+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [#4690](https://github.com/bbatsov/rubocop/issues/4690): Add new `Lint/UnneededRequireStatement` cop. ([@koic][])
 * [#4813](https://github.com/bbatsov/rubocop/pull/4813): Add new `Style/StderrPuts` cop. ([@koic][])
 * [#4796](https://github.com/bbatsov/rubocop/pull/4796): Add new `Lint/RedundantWithObject` cop. ([@koic][])
+* Add `IndentationWidth` configuration for `Layout/Tab` cop. ([@rrosenblum][])
 
 ### Bug fixes